+++ tests/Feature/Resources/WebDavAccountResourceTest.php
@@ -4,11 +4,11 @@ declare(strict_types=1);
 
 namespace N3XT0R\LaravelWebdavServerFilament\Tests\Feature\Resources;
 
-use Filament\Forms\Components\TextInput;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Str;
 use Livewire\Livewire;
 use N3XT0R\LaravelWebdavServer\Facades\WebDavPath;
+use N3XT0R\LaravelWebdavServerFilament\Filament\Forms\Components\WebDavUrlInput;
 use N3XT0R\LaravelWebdavServer\Models\WebDavAccountModel;
 use N3XT0R\LaravelWebdavServerFilament\Resources\WebDavAccountResource\Pages\CreateWebDavAccount;
 use N3XT0R\LaravelWebdavServerFilament\Resources\WebDavAccountResource\Pages\EditWebDavAccount;
@@ -363,7 +363,7 @@ final class WebDavAccountResourceTest extends DatabaseTestCase
         Livewire::test(ViewWebDavAccount::class, ['record' => $account->getKey()])
             ->assertFormFieldExists('username')
             ->assertFormFieldExists('display_name')
-            ->assertFormFieldExists('webdav_url', function (TextInput $field): bool {
+            ->assertFormFieldExists('webdav_url', function (WebDavUrlInput $field): bool {
                 self::assertTrue($field->isReadOnly());
                 self::assertTrue($field->isCopyable());